+++ Stats.h
@@ -8,18 +8,18 @@ class Stats
 {
 public:
 
-	Stats(Enemy *e, int hp, double att, double def, double wis, double intel);
+	Stats(Enemy *e, int hp, int att, int def, int wis, int intel);
 	double *getStats();
-	void change(std::vector<Buffs> b, std::vector<double> changes);
+	void change(std::vector<Buffs> b, std::vector<int> changes);
 
 private:
 	Enemy * eEnemy;
 	//Player* ePlayer;
 	int HP;
-	double att;
-	double def;
-	double wis;
-	double intel;
+	int att;
+	int def;
+	int wis;
+	int intel;
 
 };
